Added dockerfile with playground symlink
This symlink will make it easier to keep things consistent in future, so
current users can mount to playground and it will always be in the
desired location. This is important as users etc can change, and the
intention is to drop user privs so no root funny business can be used to
potentially violate docker.