Added CoC with extra bloat
So im conflicted. I have added conqueror of completion, we have some
really really nice completion, and contextual help etc. But it makes the
image quite large. I would normally use a multi stage docker build to
slim things down, but this is meant to be an IDE. If I were to then take
away the dev necessary components it would become less usefull as an
IDE, plus that would defeat the point in installing them in the first
instance. So do I A: not have the added dev stuff for CoC but have a
slim image, or B: have the
added dev stuff but live with the image size.